Push FormListFragment down into only subtype
